requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
DateTime==5.1
db-dtypes==1.1.1
google-analytics-admin==0.18.1
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pandas-gbq==0.19.2
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
zope.interface==6.0
